Moved css files into assets/styles
