+++ client/src/assets/styles/BookCoverCard.css
