+++ client/src/assets/styles/BookDetailsCard.css
